+++ src/main/java/module-info.java
@@ -10,8 +10,6 @@ module com.bumba.tic_tac_toe {
     exports com.bumba.tic_tac_toe;
     exports com.bumba.tic_tac_toe.client;
     opens com.bumba.tic_tac_toe.client to javafx.fxml;
-    exports com.bumba.tic_tac_toe.Controller;
-    opens com.bumba.tic_tac_toe.Controller to javafx.fxml;
     exports com.bumba.tic_tac_toe.server;
     opens com.bumba.tic_tac_toe.server to javafx.fxml;
     exports com.bumba.tic_tac_toe.game;
