+++ src/main/resources/com/bumba/tic_tac_toe/gameSession.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.HBox?>
 
-<HBox prefWidth="490.0" spacing="10" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.bumba.tic_tac_toe.client.Controller.LobbyController">
+<HBox prefWidth="490.0" spacing="10" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.bumba.tic_tac_toe.LobbyController">
     <Label prefHeight="26.0" prefWidth="150.0">
       <HBox.margin>
          <Insets right="20.0" />
